Mainly reorganized files and finished cleaning up.  - Set .gitignores for the maps and saves directories.  - Removed testing directory.  - Removed old hdfm script and replaced with the main execution file.  - Added a method to the main script (hdfm.py) for cleaning the dump    directory.
