+++ maps/.gitignore
@@ -1,4 +1,5 @@
 # Ignore everything in this directory.
 *
-# Except this file.
+# Except this file and the US map.
 !.gitignore
+!us_map.png
