--- testing/dump/TMI_03g65g_rev1_025d.txt
@@ -1,11 +1,0 @@
-TrafficMapProtocolVersionID="1.3"
-TrafficMapID="03g65g"
-StationList="(pIESMA,FM98.1)";"(pIEKbg,FM100.1)";"(pIEKcA,FM94.5)";"(pIBzhw,FM104.5)"
-NumRows="3"
-NumColumns="3"
-NumTransmittedTiles="9"
-CoordinatesRow1="(38.38593,-84.87529)";"(38.14798,-84.63735)";"(38.38593,-84.39940)";"(38.14798,-84.16146)"
-CoordinatesRow2="(38.14798,-84.87529)";"(37.91004,-84.63735)";"(38.14798,-84.39940)";"(37.91004,-84.16146)"
-CoordinatesRow3="(37.91004,-84.87529)";"(37.67209,-84.63735)";"(37.91004,-84.39940)";"(37.67209,-84.16146)"
-BackgroundRGBColor="(194,187,96)"
-CopyrightNotice="Copyright © 2014 iHeartMedia, Inc. All rights reserved."
